Fix: use the correct token to sign requests in the code example.
During the oauth dance two tokens pairs are generated and saved into a session;
getAccessToken() passes the most recent token/token_secret pair to the callback
function. This couple is the actual access token that needs to be saved in the current
session and has to be used to sign futher requests.

See: https://developer.linkedin.com/documents/getting-oauth-token